Move to use traits::asset_registry::AssetMetadata
The AssetMetadata type is needed by the Inspect and Mutate traits so it
needs to be in a shared place. While it's fine for the pallet to depend
on the traits::asset_registry, the inverse not so much, so we move the
struct to the traits section and import it here.